+++ test/svg/test_svg.py
@@ -280,6 +280,29 @@ class TestSVGObject:
 
         assert_pdf_equal(pdf, GENERATED_PDF_DIR / f"{svg_file.stem}.pdf", tmp_path)
 
+    @pytest.mark.parametrize(
+        "fill_color, stroke_color, file_suffix", parameters.svg_current_color
+    )
+    def test_svg_current_color(self, tmp_path, fill_color, stroke_color, file_suffix):
+        svg_file = parameters.svgfile("simple_rect_current_color.svg")
+
+        svg = fpdf.svg.SVGObject.from_file(svg_file)
+
+        pdf = fpdf.FPDF(unit="pt", format=(svg.width, svg.height))
+        pdf.set_margin(0)
+        pdf.allow_images_transparency = False
+        pdf.add_page()
+
+        if fill_color is not None:
+            pdf.set_fill_color(fill_color[0], fill_color[1], fill_color[2])
+        if stroke_color is not None:
+            pdf.set_draw_color(stroke_color[0], stroke_color[1], stroke_color[2])
+        svg.draw_to_page(pdf)
+
+        assert_pdf_equal(
+            pdf, GENERATED_PDF_DIR / f"{svg_file.stem}{file_suffix}.pdf", tmp_path
+        )
+
     def test_svg_render_content_in_a_tag(self, tmp_path):
         svg_file = parameters.svgfile("simple_rect_in_a_tag.svg")